+++ src/main/java/io/coodoo/framework/listing/boundary/ListingService.java
@@ -6,7 +6,7 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
-import io.coodoo.framework.listing.control.ListingFilterQuery;
+import io.coodoo.framework.listing.control.ListingQuery;
 
 @Stateless
 public class ListingService {
@@ -14,20 +14,20 @@ public class ListingService {
     @PersistenceContext
     EntityManager entityManager;
 
-    public <T> ListingResult<T> getListingResult(Class<T> entityClass, ListingQueryParams queryParams) {
+    public <T> ListingResult<T> getListingResult(Class<T> entityClass, ListingParameters queryParams) {
         return new ListingResult<T>(getListing(entityClass, queryParams), new Metadata(countListing(entityClass, queryParams), queryParams));
     }
 
-    public <T> List<T> getListing(Class<T> entityClass, ListingQueryParams queryParams) {
+    public <T> List<T> getListing(Class<T> entityClass, ListingParameters queryParams) {
         return assambleFilter(entityClass, queryParams).list(queryParams.getIndex(), queryParams.getLimit());
     }
 
-    public <T> Long countListing(Class<T> entityClass, ListingQueryParams queryParams) {
+    public <T> Long countListing(Class<T> entityClass, ListingParameters queryParams) {
         return assambleFilter(entityClass, queryParams).count();
     }
 
-    private <T> ListingFilterQuery<T> assambleFilter(Class<T> entityClass, ListingQueryParams queryParams) {
-        return new ListingFilterQuery<>(entityManager, entityClass)
+    private <T> ListingQuery<T> assambleFilter(Class<T> entityClass, ListingParameters queryParams) {
+        return new ListingQuery<>(entityManager, entityClass)
                         // apply sorting
                         .sort(queryParams.getSortAttribute(), queryParams.isSortAsc())
                         // disjunctive filter on the whole table
@@ -39,27 +39,27 @@ public class ListingService {
     }
 
     public <T> ListingResult<T> getListingResult(Class<T> entityClass, Integer page, Integer limit) {
-        return getListingResult(entityClass, new ListingQueryParams(page, limit, null));
+        return getListingResult(entityClass, new ListingParameters(page, limit, null));
     }
 
     public <T> List<T> getListing(Class<T> entityClass, Integer page, Integer limit) {
-        return getListing(entityClass, new ListingQueryParams(page, limit, null));
+        return getListing(entityClass, new ListingParameters(page, limit, null));
     }
 
     public <T> Long countListing(Class<T> entityClass, Integer page, Integer limit) {
-        return countListing(entityClass, new ListingQueryParams(page, limit, null));
+        return countListing(entityClass, new ListingParameters(page, limit, null));
     }
 
     public <T> ListingResult<T> getListingResult(Class<T> entityClass, Integer page, Integer limit, String sortAttribute) {
-        return getListingResult(entityClass, new ListingQueryParams(page, limit, sortAttribute));
+        return getListingResult(entityClass, new ListingParameters(page, limit, sortAttribute));
     }
 
     public <T> List<T> getListing(Class<T> entityClass, Integer page, Integer limit, String sortAttribute) {
-        return getListing(entityClass, new ListingQueryParams(page, limit, sortAttribute));
+        return getListing(entityClass, new ListingParameters(page, limit, sortAttribute));
     }
 
     public <T> Long countListing(Class<T> entityClass, Integer page, Integer limit, String sortAttribute) {
-        return countListing(entityClass, new ListingQueryParams(page, limit, sortAttribute));
+        return countListing(entityClass, new ListingParameters(page, limit, sortAttribute));
     }
 
 }